Refactor converter and reader
