+++ lib/plugin/blade/common/controller/monitor-binding-test.ts
@@ -6,8 +6,8 @@ import {MonitorBinding} from '../../../common/binding/monitor';
 import {BindingTarget} from '../../../common/binding/target';
 import {ManualTicker} from '../../../common/binding/ticker/manual';
 import {NumberFormatter} from '../../../common/converter/number';
+import {numberFromUnknown} from '../../../common/converter/number';
 import {Value} from '../../../common/model/value';
-import {numberFromUnknown} from '../../../common/reader/number';
 import {SingleLogMonitorController} from '../../../monitor-bindings/common/controller/single-log';
 import {Blade} from '../model/blade';
 import {MonitorBindingController} from './monitor-binding';
