+++ lib/plugin/common/converter/boolean.ts
@@ -6,6 +6,15 @@ import {Formatter} from './formatter';
 export function boolToString(value: boolean): string {
 	return String(value);
 }
+/**
+ * @hidden
+ */
+export function boolFromUnknown(value: unknown): boolean {
+	if (value === 'false') {
+		return false;
+	}
+	return !!value;
+}
 
 /**
  * @hidden
