+++ lib/plugin/common/converter/color-test.ts
@@ -5,6 +5,9 @@ import {Color} from '../model/color';
 import {ColorComponents3, ColorComponents4} from '../model/color-model';
 import {
 	ColorFormatter,
+	colorFromObject,
+	colorFromRgbaNumber,
+	colorFromRgbNumber,
 	colorToFunctionalHslaString,
 	colorToFunctionalHslString,
 	colorToFunctionalRgbaString,
@@ -13,9 +16,35 @@ import {
 	colorToHexRgbString,
 	colorToRgbaNumber,
 	colorToRgbNumber,
+	numberToRgbColor,
 } from './color';
 
 describe(ColorFormatter.name, () => {
+	[
+		{
+			input: {r: 0x00, g: 0x78, b: 0xff},
+			expected: {
+				r: 0x00,
+				g: 0x78,
+				b: 0xff,
+				a: 1,
+			},
+		},
+		{
+			input: {foo: 'bar'},
+			expected: {r: 0, g: 0, b: 0, a: 1},
+		},
+	].forEach((testCase) => {
+		context(`when input = ${JSON.stringify(testCase.input)}`, () => {
+			it('should convert object to color', () => {
+				assert.deepStrictEqual(
+					colorFromObject(testCase.input).toRgbaObject(),
+					testCase.expected,
+				);
+			});
+		});
+	});
+
 	it('should format color with specified stringifier', () => {
 		const stringifier = (color: Color): string => {
 			return String(color);
@@ -229,6 +258,77 @@ describe(ColorFormatter.name, () => {
 				} else {
 					throw new Error('should not be called');
 				}
+			});
+		});
+	});
+
+	[
+		{
+			expected: {r: 0x11, g: 0x22, b: 0x33, a: 1},
+			input: 0x112233,
+		},
+		{
+			expected: {r: 0x00, g: 0xaa, b: 0xff, a: 1},
+			input: 0x00aaff,
+		},
+	].forEach((testCase) => {
+		context(`when ${JSON.stringify(testCase.input)}`, () => {
+			it(`it should parse as ${JSON.stringify(testCase.expected)}`, () => {
+				const actual = numberToRgbColor(testCase.input);
+				assert.deepStrictEqual(
+					actual && actual.toRgbaObject(),
+					testCase.expected,
+				);
+			});
+		});
+	});
+
+	[
+		{
+			input: 0x0078ff,
+			expected: {
+				r: 0x00,
+				g: 0x78,
+				b: 0xff,
+				a: 1,
+			},
+		},
+		{
+			input: 'Not a number',
+			expected: {r: 0, g: 0, b: 0, a: 1},
+		},
+	].forEach((testCase) => {
+		context(`when input = ${JSON.stringify(testCase.input)}`, () => {
+			it('should convert number to rgb color', () => {
+				assert.deepStrictEqual(
+					colorFromRgbNumber(testCase.input).toRgbaObject(),
+					testCase.expected,
+				);
+			});
+		});
+	});
+
+	[
+		{
+			input: 0x9abcde33,
+			expected: {
+				r: 0x9a,
+				g: 0xbc,
+				b: 0xde,
+				a: 0.2,
+			},
+		},
+		{
+			input: 'Not a number',
+			expected: {r: 0, g: 0, b: 0, a: 1},
+		},
+	].forEach((testCase) => {
+		context(`when input = ${JSON.stringify(testCase.input)}`, () => {
+			it('should convert number to rgba color', () => {
+				assert.deepStrictEqual(
+					colorFromRgbaNumber(testCase.input).toRgbaObject(),
+					testCase.expected,
+				);
 			});
 		});
 	});
