+++ lib/plugin/common/converter/number.ts
@@ -1,4 +1,24 @@
+import {isEmpty} from '../../../misc/type-util';
+import {StringNumberParser} from '../reader/string-number';
 import {Formatter} from './formatter';
+
+/**
+ * @hidden
+ */
+export function numberFromUnknown(value: unknown): number {
+	if (typeof value === 'number') {
+		return value;
+	}
+
+	if (typeof value === 'string') {
+		const pv = StringNumberParser(value);
+		if (!isEmpty(pv)) {
+			return pv;
+		}
+	}
+
+	return 0;
+}
 
 /**
  * @hidden
