--- lib/plugin/common/reader/boolean.ts
@@ -1,9 +1,0 @@
-/**
- * @hidden
- */
-export function boolFromUnknown(value: unknown): boolean {
-	if (value === 'false') {
-		return false;
-	}
-	return !!value;
-}
