--- lib/plugin/common/reader/string.ts
@@ -1,6 +1,0 @@
-/**
- * @hidden
- */
-export function stringFromUnknown(value: unknown): string {
-	return String(value);
-}
