+++ lib/plugin/input-bindings/color/controller/color-picker.ts
@@ -1,4 +1,5 @@
 import {ValueController} from '../../../common/controller/value';
+import {NumberFormatter} from '../../../common/converter/number';
 import {findNextTarget, supportsTouch} from '../../../common/dom-util';
 import {Color} from '../../../common/model/color';
 import {Foldable} from '../../../common/model/foldable';
@@ -6,7 +7,6 @@ import {PickedColor} from '../../../common/model/picked-color';
 import {Value} from '../../../common/model/value';
 import {connect} from '../../../common/model/value-sync';
 import {StringNumberParser} from '../../../common/reader/string-number';
-import {NumberFormatter} from '../../../common/writer/number';
 import {TextController} from '../../common/controller/text';
 import {NumberTextController} from '../../number/controller/number-text';
 import {ColorPickerView} from '../view/color-picker';
