+++ lib/plugin/input-bindings/color/controller/color-swatch-text.ts
@@ -1,8 +1,8 @@
 import {ValueController} from '../../../common/controller/value';
+import {Formatter} from '../../../common/converter/formatter';
 import {Color} from '../../../common/model/color';
 import {Value} from '../../../common/model/value';
 import {Parser} from '../../../common/reader/parser';
-import {Formatter} from '../../../common/writer/formatter';
 import {TextController} from '../../common/controller/text';
 import {ColorSwatchTextView} from '../view/color-swatch-text';
 import {ColorSwatchController} from './color-swatch';
