+++ lib/plugin/input-bindings/color/view/color-component-texts.ts
@@ -1,9 +1,9 @@
+import {NumberFormatter} from '../../../common/converter/number';
 import {Color} from '../../../common/model/color';
 import {PickedColor} from '../../../common/model/picked-color';
 import {Value} from '../../../common/model/value';
 import {ClassName} from '../../../common/view/class-name';
 import {ValueView} from '../../../common/view/value';
-import {NumberFormatter} from '../../../common/writer/number';
 
 interface Config {
 	pickedColor: PickedColor;
