+++ lib/plugin/input-bindings/color/view/color-swatch.ts
@@ -1,8 +1,8 @@
+import {colorToHexRgbaString} from '../../../common/converter/color';
 import {Color} from '../../../common/model/color';
 import {Value} from '../../../common/model/value';
 import {ClassName} from '../../../common/view/class-name';
 import {ValueView} from '../../../common/view/value';
-import {colorToHexRgbaString} from '../../../common/writer/color';
 import {ColorPickerView} from './color-picker';
 
 interface Config {
