+++ lib/plugin/input-bindings/common/view/text.ts
@@ -1,7 +1,7 @@
+import {Formatter} from '../../../common/converter/formatter';
 import {Value} from '../../../common/model/value';
 import {ClassName} from '../../../common/view/class-name';
 import {ValueView} from '../../../common/view/value';
-import {Formatter} from '../../../common/writer/formatter';
 
 interface Config<T> {
 	formatter: Formatter<T>;
