+++ lib/plugin/input-bindings/number/controller/slider-text.ts
@@ -1,7 +1,7 @@
 import {ValueController} from '../../../common/controller/value';
+import {Formatter} from '../../../common/converter/formatter';
 import {Value} from '../../../common/model/value';
 import {Parser} from '../../../common/reader/parser';
-import {Formatter} from '../../../common/writer/formatter';
 import {SliderTextView} from '../view/slider-text';
 import {NumberTextController} from './number-text';
 import {SliderController} from './slider';
