+++ lib/plugin/input-bindings/point-2d/controller/point-2d-text-test.ts
@@ -2,9 +2,9 @@ import {assert} from 'chai';
 import {describe, it} from 'mocha';
 
 import {TestUtil} from '../../../../misc/test-util';
+import {NumberFormatter} from '../../../common/converter/number';
 import {Value} from '../../../common/model/value';
 import {StringNumberParser} from '../../../common/reader/string-number';
-import {NumberFormatter} from '../../../common/writer/number';
 import {Point2d} from '../model/point-2d';
 import {Point2dTextController} from './point-2d-text';
