+++ lib/plugin/input-bindings/point-3d/view/point-3d-text.ts
@@ -1,7 +1,7 @@
+import {Formatter} from '../../../common/converter/formatter';
 import {Value} from '../../../common/model/value';
 import {ClassName} from '../../../common/view/class-name';
 import {ValueView} from '../../../common/view/value';
-import {Formatter} from '../../../common/writer/formatter';
 import {Point3d} from '../model/point-3d';
 
 interface Config {
