+++ lib/plugin/monitor-bindings/common/view/multi-log.ts
@@ -1,7 +1,7 @@
+import {Formatter} from '../../../common/converter/formatter';
 import {Buffer, BufferedValue} from '../../../common/model/buffered-value';
 import {ClassName} from '../../../common/view/class-name';
 import {ValueView} from '../../../common/view/value';
-import {Formatter} from '../../../common/writer/formatter';
 
 interface Config<T> {
 	formatter: Formatter<T>;
