+++ lib/plugin/monitor-bindings/number/controller/graph-log.ts
@@ -1,8 +1,8 @@
 import {ValueController} from '../../../common/controller/value';
+import {Formatter} from '../../../common/converter/formatter';
 import {Buffer, BufferedValue} from '../../../common/model/buffered-value';
 import {GraphCursor} from '../../../common/model/graph-cursor';
 import {mapRange} from '../../../common/number-util';
-import {Formatter} from '../../../common/writer/formatter';
 import {GraphLogView} from '../view/graph-log';
 
 interface Config {
