+++ lib/plugin/monitor-bindings/number/plugin.ts
@@ -2,9 +2,9 @@ import {MonitorParams} from '../../../api/types';
 import {Constants} from '../../../misc/constants';
 import {MonitorBinding} from '../../common/binding/monitor';
 import {ValueController} from '../../common/controller/value';
+import {NumberFormatter} from '../../common/converter/number';
+import {numberFromUnknown} from '../../common/converter/number';
 import {Buffer} from '../../common/model/buffered-value';
-import {numberFromUnknown} from '../../common/reader/number';
-import {NumberFormatter} from '../../common/writer/number';
 import {MonitorBindingPlugin} from '../../monitor-binding';
 import {MultiLogController} from '../common/controller/multi-log';
 import {SingleLogMonitorController} from '../common/controller/single-log';
